requirements.txt
autopep8==1.7.0
Flask==2.2.2
Frozen-Flask==0.18
gunicorn==20.1.0
Markdown==3.4.1
Pygments==2.13.0
PyYAML==6.0
